Enabling the random volume sampler through geometric modification
Here, the geometric specifications of VT1 have been altered to allow
for proper operation of the random sampler used in the PrimaryGenerator
class. Pictures of the actual VT1 detector sent by Jim Trimble indicate
that several volumes specified as cylinders in the original
specifications are, in fact, cylindrical shells, and the code has been
changed to reflect this. Further, two volumes of air that surround the
metal components of the detector have been commented out and replaced
by a single square mother volume with X, Y, and Z half-lengths of 6m.
Lastly, GeneralParticleSource, or GPS, has been disabled, and the
random volume sampler has been put in place. Note that this code is
currently designed to provide a cobalt-60 calibration spectrum, and
that future revisions of the code will include the lead sample used in
actual testing (dimensions will be specified in a future commit).